+++ src/services/__tests__/diagnostic-updater.spec.ts
@@ -44,6 +44,7 @@ jest.mock('../../types', () => {
 });
 
 describe('DiagnosticUpdater', () => {
+	let phpcsIntegrationPath: string;
 	let mockLogger: Logger;
 	let mockWorkspaceLocator: WorkspaceLocator;
 	let mockConfiguration: Configuration;
@@ -54,15 +55,15 @@ describe('DiagnosticUpdater', () => {
 	let diagnosticUpdater: DiagnosticUpdater;
 
 	beforeEach(() => {
-		// Make sure the test knows where the real assets are located.
-		process.env.ASSETS_PATH = resolvePath(
+		phpcsIntegrationPath = resolvePath(
 			__dirname,
 			'..',
 			'..',
 			'..',
-			'assets'
+			'..',
+			'assets',
+			'phpcs-integration'
 		);
-
 		mockLogger = new Logger(window);
 		mockWorkspaceLocator = new WorkspaceLocator(workspace);
 		mockConfiguration = new Configuration(workspace, mockWorkspaceLocator);
@@ -104,6 +105,7 @@ describe('DiagnosticUpdater', () => {
 			exclude: [],
 			lintAction: LintAction.Change,
 			standard: 'PSR12',
+			phpcsIntegrationPath: phpcsIntegrationPath,
 		});
 		jest.mocked(mockWorker).execute.mockImplementation((request) => {
 			expect(request).toMatchObject({
@@ -112,6 +114,7 @@ describe('DiagnosticUpdater', () => {
 				options: {
 					executable: 'phpcs-test',
 					standard: 'PSR12',
+					phpcsIntegrationPath: phpcsIntegrationPath,
 				},
 			});
 
@@ -166,6 +169,7 @@ describe('DiagnosticUpdater', () => {
 			exclude: [],
 			lintAction: LintAction.Change,
 			standard: 'PSR12',
+			phpcsIntegrationPath: phpcsIntegrationPath,
 		});
 		jest.mocked(mockWorker).execute.mockImplementation((request) => {
 			expect(request).toMatchObject({
@@ -174,6 +178,7 @@ describe('DiagnosticUpdater', () => {
 				options: {
 					executable: 'phpcs-test',
 					standard: 'PSR12',
+					phpcsIntegrationPath: phpcsIntegrationPath,
 				},
 			});
 
@@ -198,6 +203,7 @@ describe('DiagnosticUpdater', () => {
 			exclude: [],
 			lintAction: LintAction.Save,
 			standard: 'PSR12',
+			phpcsIntegrationPath: phpcsIntegrationPath,
 		});
 
 		return diagnosticUpdater.update(document, LintAction.Change);