+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.0] - 2024-01-26
 ### Added
 - `phpCodeSniffer.specialOptions` option to allow for supporting narrow use-cases without
 needing to add options that most users will not need.